ETCM-463: Move PeerStat to a separate file.
