+++ src/main/scala/io/iohk/ethereum/network/PeerStatisticsActor.scala
@@ -1,17 +1,14 @@
 package io.iohk.ethereum.network
 
 import akka.actor._
-import cats._
-import cats.implicits._
 import io.iohk.ethereum.network.PeerEventBusActor._
 import io.iohk.ethereum.network.p2p.messages.Codes
 import java.time.Clock
 import scala.concurrent.duration.FiniteDuration
-import cats.kernel.Monoid
 
 class PeerStatisticsActor(
     peerEventBus: ActorRef,
-    var maybeStats: Option[TimeSlotStats[PeerId, PeerStatisticsActor.Stat]]
+    var maybeStats: Option[TimeSlotStats[PeerId, PeerStat]]
 ) extends Actor {
   import PeerStatisticsActor._
 
@@ -27,7 +24,7 @@ class PeerStatisticsActor(
   private def handlePeerEvents: Receive = {
     case PeerEvent.MessageFromPeer(msg, peerId) =>
       val now = System.currentTimeMillis()
-      val obs = Stat(
+      val obs = PeerStat(
         responsesReceived = if (ResponseCodes(msg.code)) 1 else 0,
         requestsReceived = if (RequestCodes(msg.code)) 1 else 0,
         firstSeenTimeMillis = Some(now),
@@ -45,49 +42,22 @@ class PeerStatisticsActor(
       sender ! StatsForAll(stats)
 
     case GetStatsForPeer(window, peerId) =>
-      val stats = maybeStats.map(_.get(peerId, Some(window))).getOrElse(Stat.empty)
+      val stats = maybeStats.map(_.get(peerId, Some(window))).getOrElse(PeerStat.empty)
       sender ! StatsForPeer(peerId, stats)
   }
 }
 
 object PeerStatisticsActor {
-  case class Stat(
-      responsesReceived: Int,
-      requestsReceived: Int,
-      firstSeenTimeMillis: Option[Long],
-      lastSeenTimeMillis: Option[Long]
-  )
-  object Stat {
-    val empty: Stat = Stat(0, 0, None, None)
-
-    private def mergeOpt[A, B](x: A, y: A)(f: A => Option[B])(g: (B, B) => B): Option[B] = {
-      val (mx, my) = (f(x), f(y))
-      (mx, my).mapN(g) orElse mx orElse my
-    }
-
-    implicit val monoid: Monoid[Stat] =
-      Monoid.instance(
-        empty,
-        (a, b) =>
-          Stat(
-            responsesReceived = a.responsesReceived + b.responsesReceived,
-            requestsReceived = a.requestsReceived + b.requestsReceived,
-            firstSeenTimeMillis = mergeOpt(a, b)(_.firstSeenTimeMillis)(math.min),
-            lastSeenTimeMillis = mergeOpt(a, b)(_.lastSeenTimeMillis)(math.max)
-          )
-      )
-  }
-
   def props(peerEventBus: ActorRef, slotDuration: FiniteDuration, slotCount: Int): Props =
     Props {
       implicit val clock = Clock.systemUTC()
-      new PeerStatisticsActor(peerEventBus, TimeSlotStats[PeerId, Stat](slotDuration, slotCount))
+      new PeerStatisticsActor(peerEventBus, TimeSlotStats[PeerId, PeerStat](slotDuration, slotCount))
     }
 
   case class GetStatsForAll(window: FiniteDuration)
-  case class StatsForAll(stats: Map[PeerId, Stat])
+  case class StatsForAll(stats: Map[PeerId, PeerStat])
   case class GetStatsForPeer(window: FiniteDuration, peerId: PeerId)
-  case class StatsForPeer(peerId: PeerId, stat: Stat)
+  case class StatsForPeer(peerId: PeerId, stat: PeerStat)
 
   val ResponseCodes = Set(
     Codes.NewBlockCode,
